separate getappstructure action to it's own file
